move campaigns post, get and patch requests functions to campaigns.js in controllers folder
