+++ my-app/controllers/campaigns.js
@@ -1,6 +1,7 @@
-const User = require("../models/User");
 const campaigns = require("../models/campaigns");
 const mongoose = require("mongoose");
+const apiKey = process.env.KEY;
+const axios = require("axios");
 
 exports.create_campaign = (req, res, next) => {
     const username = req.params.username;
@@ -19,6 +20,78 @@ exports.create_campaign = (req, res, next) => {
         res.status(200).json({
             message: 'Campaign Created!',
         });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+}
+
+exports.campaign_add_creator = (req, res, next) => {
+    const campaign = req.params.campaignName;
+    campaigns.updateOne({campaignName: campaign }, {$push:{creators: {
+        creatorId: req.body.creatorId,
+        dateJoined: new Date(),
+    }}})
+    .exec()
+    .then(result => {
+        res.status(200).json({
+            message: 'Creator added to campaign',
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+}
+
+exports.get_campaign_creators = async (req, res, next) => {
+    const username = req.params.username;
+    const campaignName = req.params.campaignName;
+    const campaignDoc = await campaigns.find({userName: username, campaignName: campaignName})
+    .then(result => {
+        //console.log(result);
+        //res.status(200);
+        return result
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+
+    //console.log(campaignDoc[0].creators);
+    const creatorsArray = campaignDoc[0].creators;
+    let creatorsIdsArray = [];
+    for(let i = 0; i < creatorsArray.length; i++){
+        creatorsIdsArray.push(creatorsArray[i].creatorId);
+    }
+    console.log(creatorsIdsArray);
+
+    const getCreators = `https://youtube.googleapis.com/youtube/v3/channels?part=snippet,topicDetails,contentDetails,statistics,brandingSettings,contentOwnerDetails&id=${creatorsIdsArray}&key=${apiKey}`;
+    let creatorsDetails = await axios.get(getCreators).then((response) => {
+        //console.log(response.data.items);
+        return response.data.items;
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+      
+    console.log(creatorsDetails);
+    res.json(creatorsDetails);
+}
+
+exports.get_campaigns = (req, res, next) => {
+    const username = req.params.username;
+    campaigns.find({userName: username})
+    .then(result => {
+        console.log(result);
+        res.status(200).json(result);
     })
     .catch(err => {
         console.log(err);
